fix(server): update handler test imports to use consolidated misc-handlers
The handler consolidation merged web-task-handlers, extension-handlers,
and environment-handlers into misc-handlers.js. Update test imports to
match the new module structure.
