+++ packages/server/tests/handlers/environment-handlers.test.js
@@ -1,6 +1,6 @@
 import { describe, it } from 'node:test'
 import assert from 'node:assert/strict'
-import { environmentHandlers } from '../../src/handlers/environment-handlers.js'
+import { miscHandlers as environmentHandlers } from '../../src/handlers/misc-handlers.js'
 import { createSpy } from '../test-helpers.js'
 
 function makeCtx(overrides = {}) {
